fix gcc and clang

+++ src/parameters/Flag.cpp
@@ -16,5 +16,5 @@ Flag::Flag(std::string &parameter) : flag(flags.find(parameter)->second){
 }
 
 Bits Flag::getBits() {
-    return {static_cast<uint32_t>(flag), 3};
+    return {static_cast<uint16_t>(flag), 3};
 }
